+++ pymatgen/io/tests/test_feffio_set.py
@@ -7,7 +7,7 @@ import os
 
 from pymatgen.io.feffio_set import FeffInputSet
 from pymatgen.io.feffio import FeffPot
-from pymatgen.io.cifio import CifParser
+from pymatgen.io.cif import CifParser
 
 test_dir = os.path.join(os.path.dirname(__file__), "..", "..", "..",
                         'test_files')
