+++ pymatgen/io/tests/test_zeoio.py
@@ -15,7 +15,7 @@ import re
 
 from pymatgen.core.periodic_table import Specie
 from pymatgen.core.structure import Structure, Molecule
-from pymatgen.io.cifio import CifParser
+from pymatgen.io.cif import CifParser
 from pymatgen.io.zeoio import ZeoCssr, ZeoVoronoiXYZ, get_voronoi_nodes, \
     get_high_accuracy_voronoi_nodes, get_void_volume_surfarea, \
     get_free_sphere_params
